Add files via upload
This 1.3 x 2.6m CNC router was designed and fitted into the bed of a pickup truck for on-site, full sheet cutting. This particular video depicts some early stage testing with a NACA1250 airfoil profile to be used as a former for a rudder on a 50 foot catamaran. Utilizes SAC controllers.